fix: ConfirmTossPaymentResponse에서 Long 타입을 Integer로 변경

+++ src/main/java/life/mosu/mosuserver/infra/payment/dto/ConfirmTossPaymentResponse.java
@@ -19,11 +19,11 @@ public class ConfirmTossPaymentResponse {
     private String orderId;
     private String status;
     private String approvedAt;
-    private Long totalAmount;
-    private Long balanceAmount;
-    private Long suppliedAmount;
-    private Long vat;
-    private Long taxFreeAmount;
+    private Integer totalAmount;
+    private Integer balanceAmount;
+    private Integer suppliedAmount;
+    private Integer vat;
+    private Integer taxFreeAmount;
     private String method;
 
     public PaymentJpaEntity toEntity(Long applicationId) {
